+++ changelog.md
@@ -39,6 +39,7 @@ Codion Change Log
 - EntityTableExport added.
 - EntityEditPanel, query inspector available with Ctrl-Alt-Q, when using a local connection, QueryInspector split into SelectQueryInspector and InsertUpdateQueryInspector.
 - SelectQueryInspector.BasicFormatterImpl bug fixed, now handles single quoted values.
+- SelectQueryInspector.BasicFormatterImpl now inserts line breaks in insert column and value lists.
 ### is.codion.swing.framework.ui.test
 - is.codion.swing.framework.ui.test module removed